Base desktop-gadget on stable upstream gadget
Drop support for rebuilding this gadget with brandstore support.

Drop building this gadget from scratch. As doing this requires to
ensure this gadget is not released to stable ahead of matching snapd &
kernel snaps. Instead stage stable gadget to always acquire stable
bootloader artefacts. This is because the primary changes in this
gadget are w.r.t. device settings.